feat: compact thumbnail position
